+++ stores/AlertStore.ts
@@ -3,18 +3,19 @@ import { create } from 'zustand';
 interface AlertStore {
   isBellDisabled: boolean;
   alertCount: number;
-  handleNewAlert: (alert: string) => void;
+  setAlertCount: (alertCount: number) => void;
+  handleNewAlert: () => void;
   handleMarkAllAsRead: () => void;
   setBellDisabled: (disabled: boolean) => void;
 }
 
 export const useAlertStore = create<AlertStore>((set) => ({
-  isBellDisabled: false,
+  isBellDisabled: true,
   alertCount: 0,
   setAlertCount: (alertCount: number) => {
     set({ alertCount: alertCount });
   },
-  handleNewAlert: (alert: string) => {
+  handleNewAlert: () => {
     set({ isBellDisabled: false });
   },
   handleMarkAllAsRead: () => {
